feat : 채팅 기능 구현

+++ src/main/java/meltingpot/server/chat/dto/ChatRoomDetailGetResponse.java
@@ -3,6 +3,6 @@ package meltingpot.server.chat.dto;
 public record ChatRoomDetailGetResponse(
         String imageKey,
         String title,
-        Long userCnt
+        Integer userCnt
 ) {
 }
